On the path to rely less on mXXXCurItem
Several changes to how listitems are managed and on how MainWindow
displays the information. Every DBRecord now owns a pointer to the
listitem referring to it. Passing the DBRecord retrieved after a click
on a listitem will make the code slicker. Also, listitems will not carry
the search array on the stack anymore. When an item is marked found,
SearchUI will ask it to load the array on the heap.