Code cleanup, bug fixes, some new features
Among the new features are two form fields that display the amount paid
in purchases for the current job and the total amount purchased. These
are calculated every time a job is displayed and their value is not
retained, even for the same session. Other noteworthy feature is the
ability to fetch the first image correponding to the day selected on
lstJobDayReport. The image selected is based on the date displayed in
its filename and if that fails (for older projects) we proceed to
searching its modification time.