Implemented the copy-and-swap idiom in some classes, notably the ones that that are overly used, such as vmWidget, VMList and its derived classes, vmTableItem, vmListItem, vmNumber, stringRecord, stringTable.
This idiom is not necessary for the current code-style used in this app,
since I do not make use of assignments, I do not reassign objects to
other objects (VMList uses pointers, and I have a mechanism to track
them and try not to screw up), etc. But, in the process of making the
classes fit for this idiom, I uncovered I few bugs, I changed the -
default - constructors to simplify them and now thoses classes are
ready, should I alter the coding style.

Some bug hunting took place, a few formatting/style/cosmetic changes.
Nothing too big