Code landed to have the mainwindow receive notifications from external (but within the program) sources about changes to the records. The code is still experimental, but the client changes (edit, add, delete) are working.
dbTableView is much more flexible. It is able to parse custom queries and
identify which types of action they will perform and what to do based on
those actions.

vivenciaDB has now two types of query runners, one that displays results
(only querys the database) and one that executes some type of change.

Eliminate all the infrastructure for custom context menus in vmWidget
and derivates. Now, context menus are handled internally via
contextMenuEvent and the widgets have a public interface to their
context menu ( standardContextMenu () ) which can be used to extend the
menu or redisplay it.

vmTableWidget displays the context menu for the cell, independently of
the widget it may or may not have.
Those changes enabled me to fix the problem of vmDateEdit not showing the
table's context menu, of the context menu overriding the widgets default
menu and of the simple tables not having a context menu

Bug fixes here and there